expense-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic-settings==2.7.1
google-api-python-client==2.159.0
google-auth==2.37.0
